requirements.txt
absl-py==2.1.0
ai4ts==0.0.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.1.0
audioread==3.0.1
beautifulsoup4==4.13.3
benchpots==0.3.2
blinker==1.9.0
cachetools==5.5.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
Crawl4AI==0.5.0.post1
cryptography==44.0.2
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==4.4.2
denoiser==0.1.5
distro==1.9.0
einops==0.8.1
et-xmlfile==1.1.0
eval_type_backport==0.2.2
exceptiongroup==1.2.0
faiss-cpu==1.10.0
fake-http-header==0.3.5
fake-useragent==2.0.3
fastapi==0.115.12
faust-cchardet==2.1.19
filelock==3.16.1
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.8.0
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanize==4.12.1
hydra-colorlog==0.1.4
hydra-core==0.11.3
idna==3.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kafka-python==2.0.2
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.59
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.6
langgraph==0.3.21
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.66
langsmith==0.3.11
lazy_loader==0.4
librosa==0.10.2.post1
litellm==1.62.1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.38.2
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.0
numpy==1.26.4; platform_machine != 'arm64'
numpy==1.26.4; platform_machine == 'arm64' --index-url https://pypi.anaconda.org/scipy-wheels-nightly/simple
omegaconf==1.4.1
openai==1.65.3
openpyxl==3.1.2
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.1
pillow==10.4.0
platformdirs==4.3.6
playwright==1.50.0
pooch==1.8.2
proglog==0.1.10
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pygrinder==0.7
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
pypots==0.9
pyreadstat==1.2.8
pyspark==3.5.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==19.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
-e git+ssh://git@github.com/bargav25/Rufus.git@40346285d048b65f486d1c27675d6ec3b667c697#egg=rufus
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.35
starlette==0.46.2
streamlit==1.45.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tf-playwright-stealth==1.1.2
threadpoolctl==3.5.0
tiktoken==0.9.0
together==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.2
transformers==4.47.1
tsdb==0.6.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
